Separate plot rendering logic in its own file
