+++ src/qlever/usage_plot.py
@@ -1,0 +1,392 @@
+from __future__ import annotations
+
+import csv
+import re
+import warnings
+from datetime import datetime
+from pathlib import Path
+
+import matplotlib
+import numpy as np
+import psutil
+
+matplotlib.use("Agg")
+from matplotlib import pyplot as plt  # noqa: E402
+
+from qlever import engine_name
+from qlever.log import log
+from qlever.resource_monitor import GB
+
+
+def read_usage_tsv(path: Path) -> dict:
+    """
+    Read a usage-log TSV into a dict of numpy arrays keyed by column
+    name. Empty cells become NaN so matplotlib can skip them.
+    """
+    columns = {}
+    with open(path) as tsv_file:
+        reader = csv.DictReader(tsv_file, delimiter="\t")
+        if reader.fieldnames is None:
+            return {}
+        for column_name in reader.fieldnames:
+            columns[column_name] = []
+        for row in reader:
+            for column_name in reader.fieldnames:
+                cell = row[column_name]
+                columns[column_name].append(float(cell) if cell else np.nan)
+    return {name: np.array(values) for name, values in columns.items()}
+
+
+def downsample_for_plot(data: dict, max_points: int) -> dict:
+    """
+    Bucket consecutive samples and reduce each bucket to one point so
+    the plot stays readable on long builds. Returns `data` unchanged
+    if it already has at most `max_points` rows.
+    """
+    n_samples = len(data["elapsed_s"])
+    if n_samples <= max_points:
+        return data
+
+    bucket_size = -(-n_samples // max_points)
+    n_buckets = -(-n_samples // bucket_size)
+    pad_length = n_buckets * bucket_size - n_samples
+
+    downsampled = {}
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", category=RuntimeWarning)
+        for column_name, series in data.items():
+            # reshape needs exact divisibility; pad the tail with NaN
+            # so the nan-aware reducers ignore the padding cells.
+            if pad_length:
+                series = np.concatenate([series, np.full(pad_length, np.nan)])
+            buckets = series.reshape(n_buckets, bucket_size)
+            # elapsed_s uses nanmin so the x-axis stays monotone (first
+            # timestamp of each bucket); rss/cpu use nanmax so peaks
+            # survive the bucketing.
+            reducer = np.nanmin if column_name == "elapsed_s" else np.nanmax
+            downsampled[column_name] = reducer(buckets, axis=1)
+    return downsampled
+
+
+def pick_time_unit(max_elapsed_s: float) -> tuple[str, float]:
+    """Pick an axis label and divisor for the X axis based on build duration."""
+    if max_elapsed_s < 200:
+        return "Elapsed (s)", 1.0
+    if max_elapsed_s < 3600:
+        return "Elapsed (min)", 60.0
+    return "Elapsed (h)", 3600.0
+
+
+def annotate_peak(
+    axes: plt.Axes,
+    x_values: np.ndarray,
+    y_values_gb: np.ndarray,
+    label: str,
+    color: str,
+    offset: tuple[int, int],
+) -> None:
+    """Draw an arrow pointing at the peak of a GB-valued series."""
+    if np.all(np.isnan(y_values_gb)):
+        return
+    peak_idx = int(np.nanargmax(y_values_gb))
+    axes.annotate(
+        f"peak {label}: {y_values_gb[peak_idx]:.2f} GB",
+        xy=(x_values[peak_idx], y_values_gb[peak_idx]),
+        xytext=offset,
+        textcoords="offset points",
+        arrowprops=dict(arrowstyle="->", color=color),
+        fontsize=9,
+        color=color,
+    )
+
+
+def compute_phase_boundaries(
+    log_path: Path,
+) -> tuple[datetime | None, dict[str, tuple[float, float]]]:
+    """
+    Parse the index build log for phase timestamps. Returns
+    (overall_begin, phases), where `overall_begin` is the datetime
+    of the first "INFO: Processing" line, and `phases` maps each
+    phase name to (start_s, end_s) elapsed seconds relative to
+    `overall_begin`. Returns (None, {}) if the log is missing or
+    the first phase cannot be located. Phases with incomplete
+    timestamps are skipped.
+    """
+    try:
+        lines = log_path.read_text().splitlines()
+    except OSError:
+        return None, {}
+
+    ts_regex = r"^(\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2})"
+    ts_format = "%Y-%m-%d %H:%M:%S"
+
+    def find(pattern: str, start: int = 0) -> tuple[datetime | None, int]:
+        for line_idx in range(start, len(lines)):
+            if re.search(pattern, lines[line_idx]):
+                match = re.match(ts_regex, lines[line_idx])
+                if match:
+                    try:
+                        return datetime.strptime(
+                            match.group(1), ts_format
+                        ), line_idx + 1
+                    except ValueError:
+                        continue
+        return None, len(lines)
+
+    overall_begin, idx = find(r"INFO:\s*Processing")
+    if overall_begin is None:
+        return None, {}
+
+    merge_begin, idx = find(r"INFO:\s*Merging partial vocab", idx)
+    convert_begin, idx = find(r"INFO:\s*Converting triples", idx)
+
+    perms = []
+    cursor = idx
+    while True:
+        perm_begin, cursor = find(
+            r"INFO:\s*Creating permutations ([A-Z]+ and [A-Z]+)", cursor
+        )
+        if perm_begin is None:
+            break
+        match = re.search(
+            r"Creating permutations ([A-Z]+ and [A-Z]+)", lines[cursor - 1]
+        )
+        name = (
+            match.group(1).replace(" and ", " & ")
+            if match
+            else f"#{len(perms) + 1}"
+        )
+        perms.append((perm_begin, name))
+
+    normal_end, idx = find(r"INFO:\s*Index build completed", idx)
+    text_begin, _ = find(r"INFO:\s*Adding text index")
+    text_end, _ = find(r"INFO:\s*Text index build comp")
+
+    def rel(ts: datetime | None) -> float | None:
+        return (ts - overall_begin).total_seconds() if ts else None
+
+    phases = {}
+
+    def add(name: str, start: datetime | None, end: datetime | None) -> None:
+        start_s, end_s = rel(start), rel(end)
+        if start_s is not None and end_s is not None:
+            phases[name] = (start_s, end_s)
+
+    add("Parse input", overall_begin, merge_begin)
+    add("Build vocabularies", merge_begin, convert_begin)
+    if perms:
+        add("Convert to global IDs", convert_begin, perms[0][0])
+        perm_names = set()
+        for perm_idx, (perm_begin, name) in enumerate(perms):
+            perm_end = (
+                perms[perm_idx + 1][0]
+                if perm_idx + 1 < len(perms)
+                else normal_end
+            )
+            if name in perm_names:
+                suffix = 2
+                while f"{name} ({suffix})" in perm_names:
+                    suffix += 1
+                name = f"{name} ({suffix})"
+            perm_names.add(name)
+            add(f"Permutation {name}", perm_begin, perm_end)
+    else:
+        add("Convert to global IDs", convert_begin, normal_end)
+    add("Text index", text_begin, text_end)
+
+    return overall_begin, phases
+
+
+def parse_git_hash(log_path: Path) -> str | None:
+    """Return the git hash printed on the first line of a QLever index log."""
+    try:
+        first_line = log_path.read_text().splitlines()[0]
+    except (OSError, IndexError):
+        return None
+    match = re.search(r"git hash ([0-9a-f]+)", first_line)
+    return match.group(1) if match else None
+
+
+def parse_qleverfile(qleverfile_path: Path) -> dict[str, str]:
+    """Read STXXL_MEMORY and num-triples-per-batch from a QLeverfile."""
+    try:
+        text = qleverfile_path.read_text()
+    except OSError:
+        return {}
+    info = {}
+    stxxl = re.search(r"^\s*STXXL_MEMORY\s*=\s*(\S+)", text, re.MULTILINE)
+    if stxxl:
+        info["stxxl"] = stxxl.group(1)
+    batch = re.search(r'"num-triples-per-batch"\s*:\s*(\d+)', text)
+    if batch:
+        triples_per_batch = int(batch.group(1))
+        if triples_per_batch >= 1_000_000:
+            info["batch"] = f"{triples_per_batch / 1_000_000:g}M"
+        elif triples_per_batch >= 1_000:
+            info["batch"] = f"{triples_per_batch / 1_000:g}K"
+        else:
+            info["batch"] = str(triples_per_batch)
+    return info
+
+
+def build_plot_subtitle(log_path: Path, qleverfile_path: Path) -> str | None:
+    """Assemble a 'batch | git | STXXL' line from the index log and QLeverfile."""
+    qleverfile_info = parse_qleverfile(qleverfile_path)
+    git_hash = parse_git_hash(log_path)
+    parts = []
+    if "batch" in qleverfile_info:
+        parts.append(f"batch = {qleverfile_info['batch']} triples")
+    if git_hash:
+        parts.append(f"git = {git_hash}")
+    if "stxxl" in qleverfile_info:
+        parts.append(f"STXXL = {qleverfile_info['stxxl']}")
+    return "   |   ".join(parts) if parts else None
+
+
+def draw_usage_plot(
+    tsv_path: Path,
+    log_path: Path,
+    qleverfile_path: Path,
+    out_path: Path,
+    title: str,
+    plot_max_points: int = 500,
+) -> None:
+    """
+    Read the usage TSV and index log, render a dual-axis plot of
+    memory and CPU over time with phase bands from the index log,
+    and save it to `out_path`. Silently returns if the TSV is empty
+    or unreadable. `plot_max_points` caps the number of dots drawn
+    per series; raw samples are bucketed and reduced with max so
+    memory peaks survive the downsampling.
+    """
+    data = read_usage_tsv(tsv_path)
+    if not data or len(data.get("elapsed_s", [])) == 0:
+        return
+
+    # drop leading rows where rss was never sampled so the plot starts
+    # at the first real measurement rather than a flat NaN run.
+    valid = np.where(~np.isnan(data["rss"]))[0]
+    if len(valid) == 0:
+        return
+    data = {name: values[valid[0] :] for name, values in data.items()}
+    data["elapsed_s"] = data["elapsed_s"] - data["elapsed_s"][0]
+
+    _, phases = compute_phase_boundaries(log_path)
+
+    elapsed_s = data["elapsed_s"]
+    if len(elapsed_s) == 0:
+        return
+
+    data = downsample_for_plot(data, plot_max_points)
+    elapsed_s = data["elapsed_s"]
+
+    x_label, x_factor = pick_time_unit(float(elapsed_s[-1]))
+    x_values = elapsed_s / x_factor
+    rss_gb = data["rss"] / GB
+    cores = psutil.cpu_count() or 1
+    cpu_cores = data["cpu_percent"] / 100.0
+
+    fig, ax_mem = plt.subplots(figsize=(12, 6), constrained_layout=True)
+    ax_cpu = ax_mem.twinx()
+
+    band_colors = plt.colormaps["Pastel1"].colors
+    total_s = float(elapsed_s[-1]) if len(elapsed_s) else 0.0
+    # skip drawing the phase name when the band is too narrow to fit it
+    # legibly; arbitrary 2% of total duration.
+    min_label_s = total_s * 0.02
+    for band_idx, (name, (start_s, end_s)) in enumerate(phases.items()):
+        band_s = end_s - start_s
+        if band_s <= 0:
+            continue
+        ax_mem.axvspan(
+            start_s / x_factor,
+            end_s / x_factor,
+            color=band_colors[band_idx % len(band_colors)],
+            alpha=0.4,
+            zorder=0,
+        )
+        if band_s < min_label_s:
+            continue
+        mid = (start_s + end_s) / 2 / x_factor
+        ax_mem.text(
+            mid,
+            0.98,
+            name,
+            transform=ax_mem.get_xaxis_transform(),
+            ha="center",
+            va="top",
+            rotation=90,
+            fontsize=8,
+            alpha=0.7,
+        )
+
+    ax_mem.plot(x_values, rss_gb, color="#cc0000", label="RSS", linewidth=1.5)
+    ax_mem.set_xlabel(x_label)
+    ax_mem.set_ylabel("RSS Memory (GB)")
+    ax_mem.grid(True, linestyle="--", alpha=0.3)
+
+    if not np.all(np.isnan(cpu_cores)):
+        ax_cpu.plot(
+            x_values,
+            cpu_cores,
+            color="#1f77b4",
+            label="CPU",
+            linewidth=1.2,
+            alpha=0.7,
+        )
+    ax_cpu.set_ylabel(f"CPU (cores, {cores} available)")
+    ax_cpu.set_ylim(0, cores)
+
+    max_rss = float(np.nanmax(rss_gb))
+    x_max = float(x_values[-1])
+    ax_mem.set_ylim(-max_rss * 0.04, max_rss * 1.4)
+    ax_mem.set_xlim(-x_max * 0.02, x_max * 1.06)
+
+    annotate_peak(ax_mem, x_values, rss_gb, "RSS", "#cc0000", (-25, 20))
+
+    lines_mem, labels_mem = ax_mem.get_legend_handles_labels()
+    lines_cpu, labels_cpu = ax_cpu.get_legend_handles_labels()
+    ax_mem.legend(
+        lines_mem + lines_cpu,
+        labels_mem + labels_cpu,
+        loc="center left",
+        bbox_to_anchor=(1.08, 0.5),
+    )
+
+    subtitle = build_plot_subtitle(log_path, qleverfile_path)
+    ax_mem.set_title(f"{title}\n{subtitle}" if subtitle else title)
+    fig.savefig(out_path, dpi=120)
+    plt.close(fig)
+
+
+def render_usage_plot(
+    dataset: str,
+    output_dir: Path | None = None,
+    plot_max_points: int = 500,
+) -> Path | None:
+    """
+    Render `<dataset>.usage-log.png` from `<dataset>.usage-log.tsv`
+    in `output_dir`. Returns the plot path on success, None if the
+    TSV is missing or the plot could not be rendered.
+    """
+    output_dir = output_dir or Path.cwd()
+    tsv_path = output_dir / f"{dataset}.usage-log.tsv"
+    log_path = output_dir / f"{dataset}.index-log.txt"
+    qleverfile_path = Path.cwd() / "Qleverfile"
+    plot_path = output_dir / f"{dataset}.usage-log.png"
+    if not tsv_path.exists():
+        log.warning(f"Resource-usage TSV not found: {tsv_path}")
+        return None
+    try:
+        draw_usage_plot(
+            tsv_path=tsv_path,
+            log_path=log_path,
+            qleverfile_path=qleverfile_path,
+            out_path=plot_path,
+            title=f"{engine_name} index build: {dataset}",
+            plot_max_points=plot_max_points,
+        )
+        return plot_path
+    except Exception as error:
+        log.warning(f"Could not render usage plot: {error}")
+        return None
